tambah midleware

+++ routes/web.php
@@ -33,19 +33,19 @@ Route::get('/login', [LoginController::class, 'index']) ->name('login') -> middl
 
 Route::post('/login', [LoginController::class, 'authenticate']);
 
-Route::get('/logout', [LoginController::class, 'logout']);
+Route::get('/logout', [LoginController::class, 'logout'])-> middleware('auth');
 
 // user
 
 Route::get('/user', [UserController::class, 'index']) -> middleware('admin');
 
-Route::post('/user', [UserController::class, 'store']);
+Route::post('/user', [UserController::class, 'store'])-> middleware('admin');
 
-Route::get('/user/{id}/edit', [UserController::class, 'edit']);
+Route::get('/user/{id}/edit', [UserController::class, 'edit'])-> middleware('admin');
 
-Route::put('/user/{id}', [UserController::class, 'update']);
+Route::put('/user/{id}', [UserController::class, 'update'])-> middleware('admin');
 
-Route::delete('user/{id}', [UserController::class, 'delete']);
+Route::delete('user/{id}', [UserController::class, 'delete'])-> middleware('admin');
 
 //content
 
@@ -57,41 +57,41 @@ Route::get('/about', [AboutController::class, 'index']) -> middleware('auth');
 
 Route::get('/spm/{spm}', [ProgramValueController::class, 'index']) -> middleware('auth');
 
-Route::get('/spm/{id}/edit', [ProgramValueController::class, 'edit']);
+Route::get('/spm/{id}/edit', [ProgramValueController::class, 'edit'])-> middleware('auth');
 
-Route::get('/spm/coba/{slug}', [ProgramValueController::class, 'show']);
+Route::get('/spm/coba/{slug}', [ProgramValueController::class, 'show'])-> middleware('auth');
 
-Route::get('/spm/filter/{program}/{year}/{id_indicator}', [ProgramValueController::class, 'filter']);
+Route::get('/spm/filter/{program}/{year}/{id_indicator}', [ProgramValueController::class, 'filter'])-> middleware('auth');
 
-Route::get('/spm/edit/{program}/{year}/{id_indicator}', [ProgramValueController::class, 'edit']);
+Route::get('/spm/edit/{program}/{year}/{id_indicator}', [ProgramValueController::class, 'edit'])-> middleware('auth');
 
-Route::put('/spm/update/{program}/{year}/{indicator}', [ProgramValueController::class, 'update']);
+Route::put('/spm/update/{program}/{year}/{indicator}', [ProgramValueController::class, 'update'])-> middleware('auth');
 
-Route::post('/indicator/{slug}', [IndicatorController::class, 'store']);
+Route::post('/indicator/{slug}', [IndicatorController::class, 'store'])-> middleware('auth');
 
 // evaluasi capaian
 
-Route::get('eval/{prog}', [EvaluasiController::class, 'index']);
+Route::get('eval/{prog}', [EvaluasiController::class, 'index'])-> middleware('auth');
 
-Route::post('eval/filter/{prog}', [EvaluasiController::class, 'filter']);
+Route::post('eval/filter/{prog}', [EvaluasiController::class, 'filter'])-> middleware('auth');
 
-Route::post('/eval/{prog}/{year}/{month}', [EvaluasiController::class, 'add']);
+Route::post('/eval/{prog}/{year}/{month}', [EvaluasiController::class, 'add'])-> middleware('auth');
 
-Route::put('/eval/{prog}', [EvaluasiController::class, 'update'])->name('eval.update');
+Route::put('/eval/{prog}', [EvaluasiController::class, 'update'])->name('eval.update')-> middleware('auth');
 
-Route::delete('eval/delete/{prog}', [EvaluasiController::class, 'delete'])->name('eval.delete');
+Route::delete('eval/delete/{prog}', [EvaluasiController::class, 'delete'])->name('eval.delete')-> middleware('auth');
 
 // report
 
-Route::get('report/{prog}', [LaporanKinerjaController::class, 'index']);
+Route::get('report/{prog}', [LaporanKinerjaController::class, 'index'])-> middleware('auth');
 
-Route::post('report/{prog}/{year}/{month}', [LaporanKinerjaController::class, 'add']);
+Route::post('report/{prog}/{year}/{month}', [LaporanKinerjaController::class, 'add'])-> middleware('auth');
 
-Route::put('report', [LaporanKinerjaController::class, 'update'])->name('report.update');
+Route::put('report', [LaporanKinerjaController::class, 'update'])->name('report.update')-> middleware('auth');
 
-Route::get('report/{prog}/{year}/{month}', [LaporanKinerjaController::class, 'filter']);
+Route::get('report/{prog}/{year}/{month}', [LaporanKinerjaController::class, 'filter'])-> middleware('auth');
 
-Route::delete('report/delete', [LaporanKinerjaController::class, 'delete'])->name('report.delete');
+Route::delete('report/delete', [LaporanKinerjaController::class, 'delete'])->name('report.delete')-> middleware('auth');
 
 // lokmun
 
@@ -109,22 +109,22 @@ Route::get('public/uploads/lokmun/{name}', [LokmunNotulenController::class, 'get
 
 // jadwal
 
-Route::get('jadwal/{prog}', [JadwalController::class, 'index']);
+Route::get('jadwal/{prog}', [JadwalController::class, 'index'])-> middleware('auth');
 
-Route::get('jadwal/{prog}/{year}/{month}', [JadwalController::class, 'filter']);
+Route::get('jadwal/{prog}/{year}/{month}', [JadwalController::class, 'filter'])-> middleware('auth');
 
-Route::post('jadwal/{prog}/{year}/{month}', [JadwalController::class, 'add']);
+Route::post('jadwal/{prog}/{year}/{month}', [JadwalController::class, 'add'])-> middleware('auth');
 
-Route::put('jadwal/{prog}//{year}/{month}', [JadwalController::class, 'update'])->name('jadwal.update');
+Route::put('jadwal/{prog}//{year}/{month}', [JadwalController::class, 'update'])->name('jadwal.update')-> middleware('auth');
 
-Route::delete('jadwal/delete/{prog}', [JadwalController::class, 'delete'])->name('jadwal.delete');
+Route::delete('jadwal/delete/{prog}', [JadwalController::class, 'delete'])->name('jadwal.delete')-> middleware('auth');
 
 // reschedule
 
-Route::get('reschedule/{prog}', [RescheduleController::class, 'index']);
+Route::get('reschedule/{prog}', [RescheduleController::class, 'index'])-> middleware('auth');
 
-Route::post('reschedule/{prog}', [RescheduleController::class, 'add']);
+Route::post('reschedule/{prog}', [RescheduleController::class, 'add'])-> middleware('auth');
 
-Route::put('reschedule/{prog}', [RescheduleController::class, 'update'])->name('reschedule.update');
+Route::put('reschedule/{prog}', [RescheduleController::class, 'update'])->name('reschedule.update')-> middleware('auth');
 
-Route::delete('reschedule/delete/{prog}', [RescheduleController::class, 'delete'])->name('reschedule.delete');
+Route::delete('reschedule/delete/{prog}', [RescheduleController::class, 'delete'])->name('reschedule.delete')-> middleware('auth');
